Authentication - Refactored utility methods in Hub class into independent clases.
